+++ src/Trixi.jl
@@ -20,17 +20,17 @@ module Trixi
 
 using LinearAlgebra: LinearAlgebra, Diagonal, diag, dot, mul!, norm, cross, normalize, I, UniformScaling
 using Printf: @printf, @sprintf, println
-using SparseArrays: AbstractSparseMatrix, AbstractSparseMatrixCSC, sparse, droptol!, rowvals, nzrange, nonzeros
+using SparseArrays: AbstractSparseMatrix, AbstractSparseMatrixCSC, sparse, droptol!, rowvals, nzrange, nonzeros, spzeros
 
 # import @reexport now to make it available for further imports/exports
 using Reexport: @reexport
 
-using DiffEqBase: @muladd, CallbackSet, DiscreteCallback,
-                  ODEProblem, ODESolution, ODEFunction
-import DiffEqBase: get_du, get_tmp_cache, u_modified!,
-                   AbstractODEIntegrator, init, step!, check_error,
-                   get_proposed_dt, set_proposed_dt!,
-                   terminate!, remake
+using SciMLBase: CallbackSet, DiscreteCallback,
+                 ODEProblem, ODESolution, ODEFunction
+import SciMLBase: get_du, get_tmp_cache, u_modified!,
+                  AbstractODEIntegrator, init, step!, check_error,
+                  get_proposed_dt, set_proposed_dt!,
+                  terminate!, remake
 using CodeTracking: code_string
 @reexport using EllipsisNotation # ..
 using ForwardDiff: ForwardDiff
@@ -40,6 +40,7 @@ using LinearMaps: LinearMap
 using LoopVectorization: LoopVectorization, @turbo, indices
 using LoopVectorization.ArrayInterface: static_length
 using MPI: MPI
+using MuladdMacro: @muladd
 using GeometryBasics: GeometryBasics
 using Octavian: Octavian, matmul!
 using Polyester: @batch # You know, the cheapest threads you can find...
@@ -50,7 +51,7 @@ using RecipesBase: RecipesBase
 using Requires: @require
 using Static: Static, One
 @reexport using StaticArrays: SVector
-using StaticArrays: MVector, MArray, SMatrix
+using StaticArrays: MVector, MArray, SMatrix, @SMatrix
 using StrideArrays: PtrArray, StrideArray, StaticInt
 @reexport using StructArrays: StructArrays, StructArray
 using TimerOutputs: TimerOutputs, @notimeit, TimerOutput, print_timer, reset_timer!
@@ -61,11 +62,13 @@ using TriplotRecipes: DGTriPseudocolor
 using UnPack: @pack!
 
 # finite difference SBP operators
-using SummationByPartsOperators: AbstractDerivativeOperator, DerivativeOperator, grid
+using SummationByPartsOperators: AbstractDerivativeOperator,
+  AbstractNonperiodicDerivativeOperator, DerivativeOperator,
+  AbstractPeriodicDerivativeOperator, PeriodicDerivativeOperator, grid
 import SummationByPartsOperators: integrate, semidiscretize,
                                   left_boundary_weight, right_boundary_weight
 @reexport using SummationByPartsOperators:
-  SummationByPartsOperators, derivative_operator
+  SummationByPartsOperators, derivative_operator, periodic_derivative_operator
 
 # DGMulti solvers
 @reexport using StartUpDG: StartUpDG, Polynomial, SBP, Line, Tri, Quad, Hex, Tet
@@ -201,7 +204,7 @@ export trixi_include, examples_dir, get_examples, default_example, default_examp
 
 export convergence_test, jacobian_fd, jacobian_ad_forward, linear_structure
 
-export DGMulti, AbstractMeshData, DGMultiMesh, estimate_dt
+export DGMulti, DGMultiMesh, estimate_dt, CartesianMesh
 export VertexMappedMesh # TODO: deprecate
 export GaussSBP